apps/api/pyproject.toml
[project]
name = "wogbuddy-backend"
version = "0.1.0"
description = "RAG Document Chat MVP"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi[standard]>=0.135.1",
    "google-genai>=1.66.0",
    "markitdown[docx,pptx,xlsx]>=0.1.5",
    "mistralai>=2.0.1",
    "openai>=2.26.0",
    "pydantic-settings>=2.13.1",
    "qdrant-client>=1.17.0",
]
